Fix Skip result behavior.
We skip current substory.  Loop will iterate methods to the first end
of story.  If the object of this bound method is one triggered Skip,
we will continue execution.

This algorithm does not take into account that story triggered Skip,
can contain substory itself.  And this second substory can be defined
in the same class.

This way we continue execution after second substory and execute wrong
piece of code.