feat: introduce class stories
Class stories are stories which are bound to the class object.
Their main purpose is to enable using stories with inheritance:

```python
from stories import class_story, Success

class Example:
  @class_story
  def x(cls, I):
    I.one

  def one(self, ctx):
    return Success()

class ExtendedExample(Example):
  @class_story
  def x(cls, I):
    super(ExtendedExample, cls).x(I)
    I.two

  def two(self, ctx):
    return Success()
```

As demonstrated by the tests, this also works with multiple inheritance.

This feature opens up more possibilities for the user
such as dynamically generating stories based on class attributes
or conditionally including a step based on a class attribute.

Implementing this feature took some refactoring
to resolve the circular references caused by importing
`wrap_story()` and `MountedStory`.

I moved `wrap_story()` to `ClassMountedStory` since it's the only place that uses it,
it has all the required state to wrap a story and it makes sense to have
a direct mapping between `ClassMountedStory` and `MountedStory`.

In addition, we don't collect strings anymore but `CallMethod` objects.
These objects allow us to customize the way we collect our steps.