feat: ignore collection of `__len__` when debugging using PyDev
When debugging story collection using PyDev (with PyCharm for example)
`__len__` is called on the `Collector` instance multiple times.
This is annoying as it interferes with debugging efforts
whenever you run the story you collected.

This commit detects that we're running using the PyDev debugger
and if that's the case, it ignores calls on `__len__`.

The code is only run when running Python without the `-O` option
which removes assert and __debug__ statements.