refactor: earthquake #627
BREAKING CHANGE: The old codebase was removed. The library was redesigned
from scratch. A lot of existed functionality was removed. Contrib integrations
was moved to separate repositories. The goal of the library stays the same.
Allow to express domain knowledge inside codebase with minimal friction. The
approach we took is very similar with the old one but most of design flaws
was fixed in new version.